add multipart test cases for openapi based clients.
