+++ lib/liff/tests/api/LiffClientTest.spec.ts
@@ -10,6 +10,8 @@ import { deepEqual, equal, ok } from "node:assert";
 
 import { describe, it } from "vitest";
 
+import { parseForm } from "../../../../test/helpers/parse-form";
+
 const channel_access_token = "test_channel_access_token";
 
 describe("LiffClient", () => {
